requirements.txt
Pillow==9.2.0
requests==2.25.1
